update happy cases

+++ view_models/transaction/transaction_request.go
@@ -1,8 +1,8 @@
 package viewmodels
 
 type TransactionRequest struct {
-	UserID        string `json:"user_id"`
-	CardID        string `json:"card_id"`
-	TransactionID string `json:"transaction_id"`
-	Amount        int64  `json:"amount"`
+	UserID        string  `json:"user_id"`
+	CardID        string  `json:"card_id"`
+	TransactionID string  `json:"transaction_id"`
+	Amount        float64 `json:"amount"`
 }
